Vss2Git

+++ Source/OldCode/Archive/2004 TVA Code Library/Source/EE/Phasor/AnalogValueBase.vb
@@ -108,7 +108,7 @@ Namespace EE.Phasor
             End Get
         End Property
 
-        Public Overrides ReadOnly Property BinaryLength() As Integer
+        Public Overrides ReadOnly Property BinaryLength() As Int16
             Get
                 If DataFormat = EE.Phasor.DataFormat.FixedInteger Then
                     Return 2
