+++ Source/OldCode/Archive/2004 TVA Code Library/Source/EE/Phasor/PhasorValueBase.vb
@@ -270,7 +270,7 @@ Namespace EE.Phasor
             End Get
         End Property
 
-        Public Overrides ReadOnly Property BinaryLength() As Integer
+        Public Overrides ReadOnly Property BinaryLength() As Int16
             Get
                 If DataFormat = EE.Phasor.DataFormat.FixedInteger Then
                     Return 4
